requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Fla  k==3.0.3
flask-marshmallow==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.2
mysql-connector-python==8.4.0
packaging==24.0
Werkzeug==3.0.3
